Delete configure package
Moved classes from configure/ to an organised fragments/ package.
