+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/BaseConfigActivity.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.mosaicnetworks.babble.configure;
+package io.mosaicnetworks.babble.fragments;
 
 import android.os.Bundle;
 import android.view.View;
@@ -32,10 +32,10 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 import io.mosaicnetworks.babble.R;
-import io.mosaicnetworks.babble.fragments.join.MdnsJoinGroupFragment;
-import io.mosaicnetworks.babble.fragments.join.WebRTCJoinGroupFragment;
 import io.mosaicnetworks.babble.fragments.archive.ArchivedGroupsFragment;
 import io.mosaicnetworks.babble.fragments.create.NewGroupFragment;
+import io.mosaicnetworks.babble.fragments.join.MdnsJoinGroupFragment;
+import io.mosaicnetworks.babble.fragments.join.WebRTCJoinGroupFragment;
 import io.mosaicnetworks.babble.fragments.tabs.TabsFragment;
 import io.mosaicnetworks.babble.servicediscovery.ResolvedGroup;
