+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/SelectableData.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.mosaicnetworks.babble.configure;
+package io.mosaicnetworks.babble.fragments;
 
 import java.util.ArrayList;
 import java.util.Collections;
