+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/archive/ArchivedGroupsFragment.java
@@ -46,8 +46,8 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import io.mosaicnetworks.babble.R;
-import io.mosaicnetworks.babble.configure.OnFragmentInteractionListener;
-import io.mosaicnetworks.babble.configure.SelectableData;
+import io.mosaicnetworks.babble.fragments.SelectableData;
+import io.mosaicnetworks.babble.fragments.OnFragmentInteractionListener;
 import io.mosaicnetworks.babble.node.ConfigDirectory;
 import io.mosaicnetworks.babble.node.ConfigManager;
 import io.mosaicnetworks.babble.service.BabbleService;
