+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/create/NewGroupFragment.java
@@ -40,8 +40,8 @@ import java.util.Objects;
 
 import androidx.annotation.NonNull;
 import io.mosaicnetworks.babble.R;
-import io.mosaicnetworks.babble.configure.BaseConfigActivity;
-import io.mosaicnetworks.babble.configure.OnFragmentInteractionListener;
+import io.mosaicnetworks.babble.fragments.BaseConfigActivity;
+import io.mosaicnetworks.babble.fragments.OnFragmentInteractionListener;
 import io.mosaicnetworks.babble.node.ConfigManager;
 import io.mosaicnetworks.babble.node.GroupDescriptor;
 import io.mosaicnetworks.babble.service.BabbleService;
